🏗️ Move component folder out of app directory
