+++ app/layout.tsx
@@ -1,9 +1,9 @@
 import { Toaster } from "@/components/ui/sonner";
 import type { Metadata, Viewport } from "next";
 import { Inter } from "next/font/google";
-import Navbar from "./components/navbar/navbar";
+import Navbar from "../components/navbar/navbar";
 import "./globals.css";
-import { ThemeProvider } from "./components/theme-provider";
+import { ThemeProvider } from "../components/theme-provider";
 import manifest from "./manifest.json";
 const inter = Inter({ subsets: ["latin"] });
