+++ components/Card.tsx
@@ -6,8 +6,8 @@ import {
 } from "@/components/ui/dialog-responsive";
 import { DialogTitle } from "@radix-ui/react-dialog";
 import Image from "next/image";
-import Info from "../detail/[slug]/info";
-import type { ItemDetails } from "../types/getDetails";
+import Info from "../app/detail/[slug]/info";
+import type { ItemDetails } from "../app/types/getDetails";
 
 export type CardProps = Pick<ItemDetails, "name" | "types" | "volumes">;
 
@@ -29,9 +29,9 @@ export default function Card({ name, types, volumes }: CardProps) {
 							alt={`${imagePath}`}
 							quality={75}
 							placeholder="blur"
-							width={300}
-							height={400}
-							className="object-cover h-full"
+							width={225}
+							height={384}
+							className="object-cover w-full h-full"
 							blurDataURL="data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
 						/>
 						{types.includes("manga") && (
