+++ components/carousel/anime/animeDetailComponent.tsx
