+++ components/carousel/manga/mangaDetailComponent.tsx
