+++ components/player.tsx
